Beta v0.8.8 (#7)
- added .gitignore
- readme changes
- added tests
- updated to version beta 0.8.8
- fixed window not staying on top
