+++ README.md
@@ -11,6 +11,7 @@ Table of contents
       * [Purpose](#purpose)
          * [What does it do?](#what-does-it-do)
          * [What <em>will</em> it do?](#what-will-it-do)
+         * [Screenshots](#screenshots)
          * [Supported platforms](#supported-platforms)
       * [Installation](#installation)
          * [Requirements](#requirements)
@@ -26,8 +27,8 @@ Table of contents
 _( generated by [gh-md-toc](https://github.com/ekalinin/github-markdown-toc) )_
 
 # Volux 
-| BRANCH  | BUILD STATUS | COVERAGE | REQUIREMENTS | ISSUES | OPEN PRs |
-| ---     | :---:        | :---:    | :---:        | :---:  | :---:    |
+| BRANCH  | BUILD | COVERAGE | REQUIREMENTS | ISSUES | OPEN PRs |
+| ---     | ---          | ---      | ---          | ---    | ---      |
 | Master  | [![Build Status](https://travis-ci.org/DrTexxOfficial/volux.svg?branch=master)](https://travis-ci.org/DrTexxOfficial/volux) | [![codecov](https://codecov.io/gh/DrTexxOfficial/volux/branch/master/graph/badge.svg)](https://codecov.io/gh/DrTexxOfficial/volux) | [![Requirements Status](https://requires.io/github/DrTexxOfficial/volux/requirements.svg?branch=master)](https://requires.io/github/DrTexxOfficial/volux/requirements/?branch=master) | [![GitHub issues](https://img.shields.io/github/issues/DrTexxOfficial/volux.svg?branch=master)](https://GitHub.com/DrTexxOfficial/volux/issues/) | [![GitHub pull-requests](https://img.shields.io/github/issues-pr/DrTexxOfficial/volux.svg?branch=master)](https://GitHub.com/DrTexxOfficial/volux/pull/) |
 | Develop | [![Build Status](https://travis-ci.org/DrTexxOfficial/volux.svg?branch=develop)](https://travis-ci.org/DrTexxOfficial/volux) | [![codecov](https://codecov.io/gh/DrTexxOfficial/volux/branch/develop/graph/badge.svg)](https://codecov.io/gh/DrTexxOfficial/volux) | [![Requirements Status](https://requires.io/github/DrTexxOfficial/volux/requirements.svg?branch=develop)](https://requires.io/github/DrTexxOfficial/volux/requirements/?branch=develop)
 
@@ -48,7 +49,14 @@ Features in development / planned for development
 - Brightness control via mouse wheel
 - Interface customisation
 
-_(this space is planned for example images of each function)_
+### Screenshots
+
+| Mode                | State    | Image |
+| :---:               | :---:    | :---: |
+| Volume (active)     | active   | <img src="docs/volume-active.jpg"/> |
+| Volume (inactive)   | inactive | <img src="docs/volume-inactive.jpg"/> |
+| Mute (active)       | active   | <img src="docs/mute-active.jpg"/> |
+| Brightness (active) | active   | <img src="docs/brightness-active.jpg"/> |
 
 ### Supported platforms
