+++ run_volux.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+cd /opt/volux && python3 -m volux
