+++ test_dp_datatools.py
@@ -1,0 +1,5 @@
+from volux import dp_datatools
+
+class TestDp_datatools():
+    def test_clamp(self):
+        assert dp_datatools.clamp(30,10,20) == 20
